+++ streamtech_salesforce_connector/models/custom_lead.py
@@ -13,4 +13,5 @@ class CustomLead(models.Model):
                                 ("Convert", "Convert"),
                                 ("Disconnected", "Disconnected"),
                                 ("Pre-Termination", "Pre-Termination"),
+                                ("Transfer", "Transfer")
                                 ], string='Type')